# openapi.json
{"openapi":"3.1.0","info":{"title":"LinguaQuest","version":"0.1.0"},"paths":{"/api/v1/auth/register":{"post":{"tags":["auth"],"summary":"Register","operationId":"register_api_v1_auth_register_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserRegister"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/TokenResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/auth/login":{"post":{"tags":["auth"],"summary":"Login","operationId":"login_api_v1_auth_login_post","requestBody":{"content":{"application/x-www-form-urlencoded":{"schema":{"$ref":"#/components/schemas/Body_login_api_v1_auth_login_post"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/TokenResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/auth/me":{"get":{"tags":["auth"],"summary":"Me","operationId":"me_api_v1_auth_me_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserProfile"}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/api/v1/auth/oauth/google/status":{"get":{"tags":["auth"],"summary":"Oauth Status","operationId":"oauth_status_api_v1_auth_oauth_google_status_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/api/v1/auth/oauth/google":{"get":{"tags":["auth"],"summary":"Oauth Google Start","operationId":"oauth_google_start_api_v1_auth_oauth_google_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/api/v1/auth/oauth/google/callback":{"get":{"tags":["auth"],"summary":"Oauth Google Callback","operationId":"oauth_google_callback_api_v1_auth_oauth_google_callback_get","parameters":[{"name":"code","in":"query","required":true,"schema":{"type":"string","title":"Code"}},{"name":"state","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"State"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/courses":{"get":{"tags":["courses"],"summary":"List Courses","operationId":"list_courses_api_v1_courses_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/api/v1/path":{"get":{"tags":["courses"],"summary":"Learning Path","operationId":"learning_path_api_v1_path_get","security":[{"OAuth2PasswordBearer":[]}],"parameters":[{"name":"course_id","in":"query","required":false,"schema":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Course Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/users/active-course":{"post":{"tags":["courses"],"summary":"Set Active Course","operationId":"set_active_course_api_v1_users_active_course_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SetActiveCourseRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/api/v1/lessons/{lesson_id}/attempts":{"post":{"tags":["lessons"],"summary":"Create Lesson Attempt","operationId":"create_lesson_attempt_api_v1_lessons__lesson_id__attempts_post","security":[{"OAuth2PasswordBearer":[]}],"parameters":[{"name":"lesson_id","in":"path","required":true,"schema":{"type":"integer","title":"Lesson Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/StartLessonResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/lesson-attempts/{attempt_id}/answer":{"post":{"tags":["lessons"],"summary":"Answer Exercise","operationId":"answer_exercise_api_v1_lesson_attempts__attempt_id__answer_post","security":[{"OAuth2PasswordBearer":[]}],"parameters":[{"name":"attempt_id","in":"path","required":true,"schema":{"type":"integer","title":"Attempt Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AnswerRequest"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/AnswerResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/lesson-attempts/{attempt_id}/complete":{"post":{"tags":["lessons"],"summary":"Finish Lesson","operationId":"finish_lesson_api_v1_lesson_attempts__attempt_id__complete_post","security":[{"OAuth2PasswordBearer":[]}],"parameters":[{"name":"attempt_id","in":"path","required":true,"schema":{"type":"integer","title":"Attempt Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/CompleteLessonResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/stats":{"get":{"tags":["gamification"],"summary":"Get Stats","operationId":"get_stats_api_v1_stats_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserStatsResponse"}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/api/v1/hearts/refill":{"post":{"tags":["gamification"],"summary":"Refill Hearts","operationId":"refill_hearts_api_v1_hearts_refill_post","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/api/v1/achievements":{"get":{"tags":["achievements"],"summary":"List Achievements","operationId":"list_achievements_api_v1_achievements_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/api/v1/leaderboard":{"get":{"tags":["achievements"],"summary":"Leaderboard","operationId":"leaderboard_api_v1_leaderboard_get","security":[{"OAuth2PasswordBearer":[]}],"parameters":[{"name":"limit","in":"query","required":false,"schema":{"type":"integer","default":20,"title":"Limit"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/friends":{"get":{"tags":["friends"],"summary":"List Friends","operationId":"list_friends_api_v1_friends_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/api/v1/friends/pending":{"get":{"tags":["friends"],"summary":"Pending Requests","operationId":"pending_requests_api_v1_friends_pending_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/api/v1/friends/request":{"post":{"tags":["friends"],"summary":"Send Request","operationId":"send_request_api_v1_friends_request_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/FriendRequestBody"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/api/v1/friends/{friendship_id}/accept":{"post":{"tags":["friends"],"summary":"Accept Request","operationId":"accept_request_api_v1_friends__friendship_id__accept_post","security":[{"OAuth2PasswordBearer":[]}],"parameters":[{"name":"friendship_id","in":"path","required":true,"schema":{"type":"integer","title":"Friendship Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/friends/{friendship_id}/reject":{"post":{"tags":["friends"],"summary":"Reject Request","operationId":"reject_request_api_v1_friends__friendship_id__reject_post","security":[{"OAuth2PasswordBearer":[]}],"parameters":[{"name":"friendship_id","in":"path","required":true,"schema":{"type":"integer","title":"Friendship Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/friends/{friendship_id}":{"delete":{"tags":["friends"],"summary":"Remove Friend","operationId":"remove_friend_api_v1_friends__friendship_id__delete","security":[{"OAuth2PasswordBearer":[]}],"parameters":[{"name":"friendship_id","in":"path","required":true,"schema":{"type":"integer","title":"Friendship Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/friends/activity":{"get":{"tags":["friends"],"summary":"Friend Activity","operationId":"friend_activity_api_v1_friends_activity_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/api/v1/recommendations":{"get":{"tags":["analytics"],"summary":"Recommendations","operationId":"recommendations_api_v1_recommendations_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/api/v1/analytics":{"get":{"tags":["analytics"],"summary":"Analytics","operationId":"analytics_api_v1_analytics_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/api/v1/analytics/learning-health":{"get":{"tags":["analytics"],"summary":"Learning Health","operationId":"learning_health_api_v1_analytics_learning_health_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/api/v1/mistakes":{"get":{"tags":["analytics"],"summary":"Mistakes","operationId":"mistakes_api_v1_mistakes_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/api/v1/practice/personalized":{"post":{"tags":["analytics"],"summary":"Personalized Practice","operationId":"personalized_practice_api_v1_practice_personalized_post","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/api/v1/practice/{session_id}/complete":{"post":{"tags":["analytics"],"summary":"Complete Practice","operationId":"complete_practice_api_v1_practice__session_id__complete_post","security":[{"OAuth2PasswordBearer":[]}],"parameters":[{"name":"session_id","in":"path","required":true,"schema":{"type":"integer","title":"Session Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/PracticeAnswer"},"title":"Answers"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/legendary/{skill_id}/start":{"post":{"tags":["analytics"],"summary":"Start Legendary","operationId":"start_legendary_api_v1_legendary__skill_id__start_post","security":[{"OAuth2PasswordBearer":[]}],"parameters":[{"name":"skill_id","in":"path","required":true,"schema":{"type":"integer","title":"Skill Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/legendary/{attempt_id}/complete":{"post":{"tags":["analytics"],"summary":"Complete Legendary","operationId":"complete_legendary_api_v1_legendary__attempt_id__complete_post","security":[{"OAuth2PasswordBearer":[]}],"parameters":[{"name":"attempt_id","in":"path","required":true,"schema":{"type":"integer","title":"Attempt Id"}},{"name":"score","in":"query","required":true,"schema":{"type":"integer","title":"Score"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/health":{"get":{"summary":"Health","operationId":"health_health_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}}},"components":{"schemas":{"AnswerRequest":{"properties":{"exercise_id":{"type":"integer","title":"Exercise Id"},"user_answer":{"title":"User Answer"}},"type":"object","required":["exercise_id","user_answer"],"title":"AnswerRequest"},"AnswerResponse":{"properties":{"is_correct":{"type":"boolean","title":"Is Correct"},"correct_answer":{"title":"Correct Answer"},"hearts_remaining":{"type":"integer","title":"Hearts Remaining"},"lesson_failed":{"type":"boolean","title":"Lesson Failed"},"feedback":{"type":"string","title":"Feedback"},"exercises_remaining":{"type":"integer","title":"Exercises Remaining"}},"type":"object","required":["is_correct","hearts_remaining","lesson_failed","feedback","exercises_remaining"],"title":"AnswerResponse"},"Body_login_api_v1_auth_login_post":{"properties":{"grant_type":{"anyOf":[{"type":"string","pattern":"password"},{"type":"null"}],"title":"Grant Type"},"username":{"type":"string","title":"Username"},"password":{"type":"string","title":"Password"},"scope":{"type":"string","title":"Scope","default":""},"client_id":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Client Id"},"client_secret":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Client Secret"}},"type":"object","required":["username","password"],"title":"Body_login_api_v1_auth_login_post"},"CompleteLessonResponse":{"properties":{"xp_earned":{"type":"integer","title":"Xp Earned"},"total_xp":{"type":"integer","title":"Total Xp"},"gems_earned":{"type":"integer","title":"Gems Earned"},"skill_progress":{"type":"object","title":"Skill Progress"},"crown_level":{"type":"integer","title":"Crown Level"},"unlocked_skills":{"items":{"type":"integer"},"type":"array","title":"Unlocked Skills"},"streak":{"type":"integer","title":"Streak"},"daily_goal":{"type":"object","title":"Daily Goal"},"newly_earned_achievements":{"items":{"type":"object"},"type":"array","title":"Newly Earned Achievements"},"perfect_lesson":{"type":"boolean","title":"Perfect Lesson"}},"type":"object","required":["xp_earned","total_xp","gems_earned","skill_progress","crown_level","unlocked_skills","streak","daily_goal","newly_earned_achievements","perfect_lesson"],"title":"CompleteLessonResponse"},"ExercisePublic":{"properties":{"id":{"type":"integer","title":"Id"},"order":{"type":"integer","title":"Order"},"type":{"type":"string","title":"Type"},"prompt":{"type":"string","title":"Prompt"},"payload":{"type":"object","title":"Payload"},"audio_text":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Audio Text"},"language":{"type":"string","title":"Language"}},"type":"object","required":["id","order","type","prompt","payload","language"],"title":"ExercisePublic","description":"Exercise without correct answers — safe for client before submission."},"FriendRequestBody":{"properties":{"user_id":{"type":"integer","title":"User Id"}},"type":"object","required":["user_id"],"title":"FriendRequestBody"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"PracticeAnswer":{"properties":{"exercise_id":{"type":"integer","title":"Exercise Id"},"user_answer":{"title":"User Answer"}},"type":"object","required":["exercise_id","user_answer"],"title":"PracticeAnswer"},"SetActiveCourseRequest":{"properties":{"course_id":{"type":"integer","title":"Course Id"}},"type":"object","required":["course_id"],"title":"SetActiveCourseRequest"},"StartLessonResponse":{"properties":{"attempt_id":{"type":"integer","title":"Attempt Id"},"lesson_id":{"type":"integer","title":"Lesson Id"},"exercises":{"items":{"$ref":"#/components/schemas/ExercisePublic"},"type":"array","title":"Exercises"},"hearts_remaining":{"type":"integer","title":"Hearts Remaining"},"total_exercises":{"type":"integer","title":"Total Exercises"}},"type":"object","required":["attempt_id","lesson_id","exercises","hearts_remaining","total_exercises"],"title":"StartLessonResponse"},"TokenResponse":{"properties":{"access_token":{"type":"string","title":"Access Token"},"refresh_token":{"type":"string","title":"Refresh Token"},"token_type":{"type":"string","title":"Token Type","default":"bearer"}},"type":"object","required":["access_token","refresh_token"],"title":"TokenResponse"},"UserProfile":{"properties":{"id":{"type":"integer","title":"Id"},"email":{"type":"string","format":"email","title":"Email"},"username":{"type":"string","title":"Username"},"display_name":{"type":"string","title":"Display Name"},"avatar_url":{"type":"string","title":"Avatar Url"},"active_course_id":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Active Course Id"},"total_xp":{"type":"integer","title":"Total Xp","default":0},"gems":{"type":"integer","title":"Gems","default":0},"hearts":{"type":"integer","title":"Hearts","default":0},"max_hearts":{"type":"integer","title":"Max Hearts","default":5},"current_streak":{"type":"integer","title":"Current Streak","default":0},"longest_streak":{"type":"integer","title":"Longest Streak","default":0},"daily_xp_goal":{"type":"integer","title":"Daily Xp Goal","default":50},"today_xp":{"type":"integer","title":"Today Xp","default":0},"lessons_completed":{"type":"integer","title":"Lessons Completed","default":0}},"type":"object","required":["id","email","username","display_name","avatar_url"],"title":"UserProfile"},"UserRegister":{"properties":{"email":{"type":"string","format":"email","title":"Email"},"username":{"type":"string","maxLength":50,"minLength":3,"title":"Username"},"password":{"type":"string","maxLength":128,"minLength":8,"title":"Password"},"display_name":{"type":"string","maxLength":100,"minLength":1,"title":"Display Name"}},"type":"object","required":["email","username","password","display_name"],"title":"UserRegister"},"UserStatsResponse":{"properties":{"total_xp":{"type":"integer","title":"Total Xp"},"gems":{"type":"integer","title":"Gems"},"hearts":{"type":"integer","title":"Hearts"},"max_hearts":{"type":"integer","title":"Max Hearts"},"current_streak":{"type":"integer","title":"Current Streak"},"longest_streak":{"type":"integer","title":"Longest Streak"},"daily_xp_goal":{"type":"integer","title":"Daily Xp Goal"},"today_xp":{"type":"integer","title":"Today Xp"},"daily_goal_met":{"type":"boolean","title":"Daily Goal Met"},"lessons_completed":{"type":"integer","title":"Lessons Completed"},"perfect_lessons":{"type":"integer","title":"Perfect Lessons"}},"type":"object","required":["total_xp","gems","hearts","max_hearts","current_streak","longest_streak","daily_xp_goal","today_xp","daily_goal_met","lessons_completed","perfect_lessons"],"title":"UserStatsResponse"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}},"securitySchemes":{"OAuth2PasswordBearer":{"type":"oauth2","flows":{"password":{"scopes":{},"tokenUrl":"/api/v1/auth/login"}}}}}}